requirements.txt
# USGS Flow Percentile Monitor - Dependencies
# TICKET-1.2: Python Environment Setup

# USGS Data Access
dataretrieval>=1.0.0
hyswap>=0.1.0

# Data Manipulation
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=14.0.0

# AWS SDK
boto3>=1.34.0

# Parallelization & Progress
tqdm>=4.66.0

# Configuration
python-dotenv>=1.0.0
pyyaml>=6.0.0

# Development & Testing
pytest>=7.4.0
pytest-cov>=4.1.0
